Check in android-current as android-26 for O.
The O preview branch didn't actually generate an android-26 (or even
an android-25), so just check in the android-current stuff. The only
extra things being exposed by doing this are some things that might
be in LIBC_FUTURE, but that's not really a problem.

These are from the same build as the rest of the O API updates:
3821160.

Test: ndk/checkbuild.py
Bug: None
Change-Id: Ib82a14b35fa811e5a803e425884be043cb08c1fa
(cherry picked from commit 9786496e836e6d181cc583ce857e6c8dce7bbe90)