Remove .llvm_addrsig section from CRT objects.
The CRT objects in NDK r20 were built with -faddrsig because that's
was the default behavior for Android, but icf=safe (which the platform
uses is incompatible with ld -r objects that were built with
-faddrsig).

We don't have many good options for fixing this short of rebuilding
NDK r20, and that's overkill. Instead, just strip the problematic
section.

for obj in $(find -name '*.o'); do
    prebuilts/clang/host/linux-x86/clang-r353983d/bin/llvm-objcopy \
        --remove-section .llvm_addrsig $obj
done

Test: make checkbuild
Bug: None
Change-Id: I4430279e4f90570c9ccc8d25fcac1ad003d5cf88