Check in GDB/gdbserver prebuilts.
Doesn't built with the next Clang update because our expat is too old,
and our build servers lack the dependencies for building a newer
expat. We don't expect to update GDB again anyway, so just check in
the binaries from r22 so we can continue shipping those without
needing to rebuild.

If a rebuild is needed, these are from the final NDK r22 release,
which was build 7026061 of aosp/ndk-release-r22, tagged as ndk-r22.
That tag can be used to rebuild these binaries (or, more easily, by
syncing using manifest_7026061.xml from the build).

https://ci.android.com/builds/branches/aosp-ndk-release-r22/grid?head=7026061&tail=7026061

Test: ndk/checkbuild.py --system windows64
Bug: None
Change-Id: I6d742b4d04d57bea7561490bdfcc9a298bcc6676