Check in prebuilt Python.
As with GDB, we don't need to build this any more. The only reason we
still *have* Python 2 is for use with GDB, so it makes sense to freeze
both if we need to freeze GDB.

If a rebuild is needed, these are from the final NDK r22 release,
which was build 7026061 of aosp/ndk-release-r22, tagged as ndk-r22.
That tag can be used to rebuild these binaries (or, more easily, by
syncing using manifest_7026061.xml from the build).

https://ci.android.com/builds/branches/aosp-ndk-release-r22/grid?head=7026061&tail=7026061

Test: ndk/checkbuild.py
Bug: None
Change-Id: I0ff8b8fb461d202808b25b1a7787f366ac8741a2