Hack in android/NeuralNetworksTypes.h.
The path that contains this header in the platform was directly
included by the CTS test that should have caught this.

This header is copied from the same revision of
//packages/modules/neuralnetworks that the rest of the artifacts came
from, so this is identical to the output that we *would* have if the
ndk_headers rule was correct.

Test: treehugger
Bug: http://b/184286703
Change-Id: I96bbe3be26b2dba4c0d84a036625ac5c48890a37
(cherry picked from commit 6616c0487cf11d3977fba7e6fe36d80ee9eed2f6)