Reapply hack in android/NeuralNetworksTypes.h.
The problem was caught in S DP 3 after the cutoff date for S beta 1
changes, so this needs to be reapplied.

This header is copied from the same revision of
//packages/modules/neuralnetworks that the rest of the artifacts came
from, so this is identical to the output that we *would* have if the
ndk_headers rule was correct.

Test: treehugger
Bug: http://b/187078104
Change-Id: I349f20f919c4e0c765bdf06c1bad2c20963dee75
(cherry picked from commit 849be7280ed95e644c13161c5100c89eee953958)