Add rv64 prebuilts directly from the toolchain.
These are not the same version of libc++ as the rest of the prebuilts,
but we need to get a V sysroot with rv64 into the NDK sooner rather
than later. We've been unable to do that because the workaround for
getting a libc++ NDK prebuilt for rv64 thus far has been
`sdk_version: "current"` in external/libcxx, but we can't set it to
"VanillaIceCream" because of trunk stable problems. This hack probably
gets us across that hurdle.

Bug: None
Test: treehugger
Change-Id: I52c291f55c56f92320cbc512b76d462068d8e9a3